+++ packages/placement-ordering/configure/src/design.jsx
@@ -1,9 +1,7 @@
 import {
   FeedbackConfig,
   FormSection,
-  InputCheckbox,
   InputContainer,
-  TwoChoice
 } from '@pie-lib/config-ui';
 import EditableHtml from '@pie-lib/editable-html';
 
@@ -12,31 +10,30 @@ import { get, set } from 'nested-property';
 import ChoiceEditor from './choice-editor';
 import PropTypes from 'prop-types';
 import React from 'react';
-import TextField from '@material-ui/core/TextField';
 import cloneDeep from 'lodash/cloneDeep';
 import debug from 'debug';
 import { withStyles } from '@material-ui/core/styles';
+import getSideMenuItems from './settings';
 
 const log = debug('@pie-element:placement-ordering:design');
 
-class Design extends React.Component {
+export class Design extends React.Component {
   constructor(props) {
     super(props);
 
-    this.state = {
-      allMoveOnDrag: false
-    };
-
     this.applyUpdate = modelFn => {
-      const { model, onModelChange } = this.props;
+      const { model, updateModel } = this.props;
       const update = modelFn(cloneDeep(model));
-      onModelChange(update);
+
+      updateModel(update);
     };
 
     this.changeHandler = (modelPath, valuePath) => {
       return value => {
         log('[changeHandler] value: ', value);
+
         const v = valuePath ? get(value, valuePath) : value;
+
         this.applyUpdate(model => {
           set(model, modelPath, v);
           return model;
@@ -44,29 +41,7 @@ class Design extends React.Component {
       };
     };
 
-    this.onLayoutChange = layout => {
-      this.applyUpdate(model => {
-        model.choiceAreaLayout = layout;
-        return model;
-      });
-    };
-
-    this.onPlacementTypeChange = event => {
-      const includePlacment = event.currentTarget.checked;
-      this.applyUpdate(model => {
-        model.placementType = includePlacment ? 'placement' : 'none';
-        return model;
-      });
-    };
-
-    this.onDefaultLangChange = defaultLang => {
-      this.applyUpdate(model => {
-        model.defaultLang = defaultLang;
-        return model;
-      });
-    };
-
-    this.onPromptChange = this.changeHandler('prompt');
+    this.onPromptChange = this.changeHandler('itemStem');
     this.onChoiceAreaLabelChange = this.changeHandler(
       'choiceAreaLabel',
       'target.value'
@@ -76,145 +51,93 @@ class Design extends React.Component {
       'target.value'
     );
     this.onFeedbackChange = this.changeHandler('feedback');
-    this.onShuffleChange = this.changeHandler('shuffle', 'target.checked');
-    this.onShowOrderingChange = this.changeHandler(
-      'showOrdering',
-      'target.checked'
-    );
 
     this.onChoiceEditorChange = (choices, correctResponse) => {
-      const { model, onModelChange } = this.props;
+      const { model, updateModel } = this.props;
       const update = cloneDeep(model);
+
       update.choices = choices;
       update.correctResponse = correctResponse;
-      onModelChange(update);
+      updateModel(update);
     };
   }
 
   render() {
     const { model, classes, imageSupport } = this.props;
-    const { configure: {
-      orientationLabel,
-      shuffleLabel,
-      includePlacementAreaLabel,
-      numberedGuidesLabel,
-      promptLabel,
-      choiceLabel,
-      choicesLabel,
-      enableOrientationChange,
-      enableShuffleChange,
-      enablePlacementAreaChange,
-      enableNumberedGuideChange,
-      enablePromptChange,
-      enableChoiceLabelChange,
-      enableChoicesLabelChange,
-      enableFeedback,
-      removeTilesLabel,
-      enableRemoveTiles
-    } } = model;
-    const { allMoveOnDrag } = this.state;
+    const {
+      configure: {
+        itemStemLabel,
+        choiceLabel,
+        choicesLabel,
+        settingsItemStemChange,
+        settingsChoicesLabel,
+        settingsFeedback,
+        settingsPlacementAreaLabel,
+        answerLabel,
+      },
+    } = model;
 
     return (
       <div className={classes.design}>
-        {
-          enableOrientationChange &&
-          <div className={classes.row}>
-            <TwoChoice
-              className={classes.orientation}
-              header={orientationLabel}
-              value={model.choiceAreaLayout}
-              onChange={this.onLayoutChange}
-              one={{ label: 'vertical', value: 'vertical' }}
-              two={{ label: 'horizontal', value: 'horizontal' }}
-            />
-          </div>
-        }
-        <div className={classes.row}>
-          {
-            enableShuffleChange &&
-            <InputCheckbox
-              label={shuffleLabel}
-              checked={model.shuffle}
-              onChange={this.onShuffleChange}
-              aria-label="shuffle"
-            />
-          }
-
-          {
-            enablePlacementAreaChange &&
-            <InputCheckbox
-              label={includePlacementAreaLabel}
-              checked={model.placementType === 'placement'}
-              onChange={this.onPlacementTypeChange}
-              aria-label="include-placment"
-            />
-          }
-
-          {
-            enableNumberedGuideChange &&
-            <InputCheckbox
-              disabled={model.placementType !== 'placement'}
-              label={numberedGuidesLabel}
-              checked={model.showOrdering}
-              onChange={this.onShowOrderingChange}
-              aria-label="shuffle"
-            />
-          }
-
+        <div className={classes.settings}>
+          {getSideMenuItems(this.props)}
         </div>
 
         {
-          enablePromptChange &&
-          <InputContainer label={promptLabel} className={classes.promptHolder}>
-            <EditableHtml
-              className={classes.prompt}
-              markup={model.prompt}
-              onChange={this.onPromptChange}
-              imageSupport={imageSupport}
-            />
-          </InputContainer>
-        }
-
-        <div className={classes.row}>
-          {
-            enableChoiceLabelChange &&
-            <TextField
-              className={classes.choiceLabel}
-              label={choiceLabel}
-              value={model.choiceAreaLabel}
-              onChange={this.onChoiceAreaLabelChange}
-              fullWidth
-            />
-          }
-
-          {model.placementType === 'placement' && (
-            <TextField
-              label="Answer label"
-              value={model.answerAreaLabel}
-              onChange={this.onAnswerAreaLabelChange}
-              fullWidth
-            />
-          )}
-        </div>
-
-        {
-          enableChoicesLabelChange &&
-          <FormSection label={choicesLabel}>
-            <ChoiceEditor
-              correctResponse={model.correctResponse}
-              choices={model.choices}
-              configuration={{
-                removeTilesLabel,
-                enableRemoveTiles,
-              }}
-              onChange={this.onChoiceEditorChange}
-              imageSupport={imageSupport}
-            />
+          settingsItemStemChange &&
+          <FormSection label="Ordering">
+            <InputContainer label={itemStemLabel && itemStemLabel.toUpperCase()} className={classes.promptHolder}>
+              <EditableHtml
+                className={classes.prompt}
+                markup={model.itemStem}
+                onChange={this.onPromptChange}
+                imageSupport={imageSupport}
+              />
+            </InputContainer>
           </FormSection>
         }
 
+        <FormSection label="Define Choices">
+          <div className={classes.row}>
+            {
+              model.configure.editableChoiceLabel && (
+                <InputContainer label={choiceLabel && choiceLabel.toUpperCase()} className={classes.promptHolder}>
+                  <EditableHtml
+                    className={classes.prompt}
+                    markup={model.choiceAreaLabel}
+                    onChange={this.onChoiceAreaLabelChange}
+                  />
+                </InputContainer>
+              )}
+
+            {(settingsPlacementAreaLabel && model.placementArea === true) && (
+              <InputContainer label={answerLabel && answerLabel.toUpperCase()} className={classes.promptHolder}>
+                <EditableHtml
+                  className={classes.prompt}
+                  markup={model.answerAreaLabel}
+                  onChange={this.onAnswerAreaLabelChange}
+                />
+              </InputContainer>
+            )}
+          </div>
+
+          {
+            settingsChoicesLabel &&
+            <InputContainer label={choicesLabel && choicesLabel.toUpperCase()} className={classes.promptHolder}>
+              <ChoiceEditor
+                correctResponse={model.correctResponse}
+                choices={model.choices}
+                onChange={this.onChoiceEditorChange}
+                imageSupport={imageSupport}
+                disableImages={model.configure && !model.configure.imagesEnabled}
+              />
+            </InputContainer>
+          }
+
+        </FormSection>
+
         {
-          enableFeedback &&
+          settingsFeedback &&
           <FeedbackConfig
             feedback={model.feedback}
             onChange={this.onFeedbackChange}
@@ -229,12 +152,16 @@ class Design extends React.Component {
 
 Design.propTypes = {
   model: PropTypes.object.isRequired,
-  onModelChange: PropTypes.func.isRequired
+  updateModel: PropTypes.func,
+  classes: PropTypes.object.isRequired,
+  imageSupport: PropTypes.object
 };
 
 export default withStyles(theme => ({
   promptHolder: {
-    width: '100%'
+    width: '100%',
+    paddingTop: '12px',
+    marginTop: '24px'
   },
   prompt: {
     paddingTop: theme.spacing.unit * 2,
@@ -260,5 +187,9 @@ export default withStyles(theme => ({
   orientation: {
     marginTop: '0px',
     marginBottom: '0px'
+  },
+  settings: {
+    display: 'flex',
+    flexWrap: 'wrap'
   }
 }))(Design);